Refactor Express Request type definitions in middleware and response utilities
- Consolidated the Request type extensions for `id` and `auth` properties into `responseUtils.ts` for better organization and maintainability.
