+++ backend/src/middleware/requestId.ts
@@ -1,15 +1,6 @@
 import { Request, Response, NextFunction } from "express";
 import { v4 as uuidv4 } from "uuid";
 
-// Extend Express Request type to include id
-declare global {
-  namespace Express {
-    interface Request {
-      id: string;
-    }
-  }
-}
-
 export const requestIdMiddleware = (
   req: Request,
   res: Response,
